Remove dependency from lxml from the main module
By using get_version_string from gvm/utils.py in gvm/__init__.py I did
accidentially introduce the dependency on lxml when running the setup.py
file again. Therefore remove the lxml dependency from gvm.utils by
moving the pretty_print function to the gvm.xml module.
