+++ docs/api.rst
@@ -80,3 +80,9 @@ Utils
 
 .. automodule:: gvm.utils
     :members:
+
+XML
+---
+
+.. automodule:: gvm.xml
+    :members:
